Refactor: build_schema_payload を resources.py へ移動 (#184)
mcp_contract.py は「tool 契約 (schema 集約)」に専念し、Pydantic モデルを
wire 形式 dict へ serialize する runtime payload 構築は新 module
``resources.py`` に分離する。

- src/vault_search/resources.py: 新規。build_schema_payload を移動
- src/vault_search/mcp_contract.py: build_schema_payload 定義と
  Iterable / FrontmatterKeyInfo import を削除。__all__ も更新
- src/vault_search/server.py: import 経路を更新
- tests: 3 ファイルの import を更新

振る舞い変更なし (純 refactor)。524 tests 全通過。

後続 PR 1b #38 で overview / recommended_flow / errors / FrontmatterKeyInfo
JSON Schema 公開を追加する際、resources.py 側で payload 合成するため
mcp_contract.py が「純 schema 生成」のまま維持できる。
